+++ src/app/cherry-picked-green/herramientas/[slug]/page.tsx
@@ -38,6 +38,7 @@ export default async function HerramientaCpPage({
         src={h.src}
         veredicto={h.veredicto}
         soportaMemoria={h.soportaMemoria}
+        guia={h.guia}
         pantallaCompleta
       />
     </main>